chore: Add OpenAPI support for the Rocket.Chat e2e endpoints
- migrating to a modern chained route definition syntax and utilizing shared AJV schemas for validation to enhance API documentation and ensure type safety through response validation.
