+++ apps/meteor/app/api/server/v1/e2e.ts
@@ -1,13 +1,11 @@
-import type { IUser } from '@rocket.chat/core-typings';
+import type { IRoom, ISubscription, IUser } from '@rocket.chat/core-typings';
 import { Subscriptions, Users } from '@rocket.chat/models';
 import {
 	ajv,
 	validateUnauthorizedErrorResponse,
 	validateBadRequestErrorResponse,
+	validateForbiddenErrorResponse,
 	ise2eSetUserPublicAndPrivateKeysParamsPOST,
-	isE2EProvideUsersGroupKeyProps,
-	isE2EFetchUsersWaitingForGroupKeyProps,
-	isE2EResetRoomKeyProps,
 } from '@rocket.chat/rest-typings';
 import ExpiryMap from 'expiry-map';
 
@@ -41,6 +39,18 @@ type e2eUpdateGroupKeyParamsPOST = {
 	uid: string;
 	rid: string;
 	key: string;
+};
+
+type E2EFetchUsersWaitingForGroupKeyProps = { roomIds: string[] };
+
+type E2EProvideUsersGroupKeyProps = {
+	usersSuggestedGroupKeys: Record<IRoom['_id'], { _id: IUser['_id']; key: string; oldKeys: ISubscription['suggestedOldRoomKeys'] }[]>;
+};
+
+type E2EResetRoomKeyProps = {
+	rid: string;
+	e2eKey: string;
+	e2eKeyId: string;
 };
 
 const E2eSetRoomKeyIdSchema = {
@@ -85,6 +95,67 @@ const e2eUpdateGroupKeyParamsPOSTSchema = {
 	required: ['uid', 'rid', 'key'],
 };
 
+const E2EFetchUsersWaitingForGroupKeySchema = {
+	type: 'object',
+	properties: {
+		roomIds: {
+			type: 'array',
+			items: {
+				type: 'string',
+			},
+		},
+	},
+	required: ['roomIds'],
+	additionalProperties: false,
+};
+
+const E2EProvideUsersGroupKeySchema = {
+	type: 'object',
+	properties: {
+		usersSuggestedGroupKeys: {
+			type: 'object',
+			additionalProperties: {
+				type: 'array',
+				items: {
+					type: 'object',
+					properties: {
+						_id: { type: 'string' },
+						key: { type: 'string' },
+						oldKeys: {
+							type: 'array',
+							items: {
+								type: 'object',
+								properties: { e2eKeyId: { type: 'string' }, ts: { type: 'string' }, E2EKey: { type: 'string' } },
+							},
+						},
+					},
+					required: ['_id', 'key'],
+					additionalProperties: false,
+				},
+			},
+		},
+	},
+	required: ['usersSuggestedGroupKeys'],
+	additionalProperties: false,
+};
+
+const E2EResetRoomKeySchema = {
+	type: 'object',
+	properties: {
+		rid: {
+			type: 'string',
+		},
+		e2eKey: {
+			type: 'string',
+		},
+		e2eKeyId: {
+			type: 'string',
+		},
+	},
+	required: ['rid', 'e2eKey', 'e2eKeyId'],
+	additionalProperties: false,
+};
+
 const isE2eSetRoomKeyIdProps = ajv.compile<E2eSetRoomKeyIdProps>(E2eSetRoomKeyIdSchema);
 
 const ise2eGetUsersOfRoomWithoutKeyParamsGET = ajv.compile<e2eGetUsersOfRoomWithoutKeyParamsGET>(
@@ -92,6 +163,12 @@ const ise2eGetUsersOfRoomWithoutKeyParamsGET = ajv.compile<e2eGetUsersOfRoomWith
 );
 
 const ise2eUpdateGroupKeyParamsPOST = ajv.compile<e2eUpdateGroupKeyParamsPOST>(e2eUpdateGroupKeyParamsPOSTSchema);
+
+const isE2EFetchUsersWaitingForGroupKeyProps = ajv.compile<E2EFetchUsersWaitingForGroupKeyProps>(E2EFetchUsersWaitingForGroupKeySchema);
+
+const isE2EProvideUsersGroupKeyProps = ajv.compile<E2EProvideUsersGroupKeyProps>(E2EProvideUsersGroupKeySchema);
+
+const isE2EResetRoomKeyProps = ajv.compile<E2EResetRoomKeyProps>(E2EResetRoomKeySchema);
 
 const e2eEndpoints = API.v1
 	.post(
@@ -191,6 +268,108 @@ const e2eEndpoints = API.v1
 		},
 	)
 	.post(
+		'e2e.rejectSuggestedGroupKey',
+		{
+			authRequired: true,
+			body: ise2eGetUsersOfRoomWithoutKeyParamsGET,
+			response: {
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+				200: ajv.compile<void>({
+					type: 'object',
+					properties: {
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['success'],
+				}),
+			},
+		},
+
+		async function action() {
+			const { rid } = this.bodyParams;
+
+			await handleSuggestedGroupKey('reject', rid, this.userId, 'e2e.rejectSuggestedGroupKey');
+
+			return API.v1.success();
+		},
+	)
+	.post(
+		'e2e.acceptSuggestedGroupKey',
+		{
+			authRequired: true,
+			body: ise2eGetUsersOfRoomWithoutKeyParamsGET,
+			response: {
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+				200: ajv.compile<void>({
+					type: 'object',
+					properties: {
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['success'],
+				}),
+			},
+		},
+
+		async function action() {
+			const { rid } = this.bodyParams;
+
+			await handleSuggestedGroupKey('accept', rid, this.userId, 'e2e.acceptSuggestedGroupKey');
+
+			return API.v1.success();
+		},
+	)
+	.get(
+		'e2e.fetchUsersWaitingForGroupKey',
+		{
+			authRequired: true,
+			query: isE2EFetchUsersWaitingForGroupKeyProps,
+			response: {
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+				200: ajv.compile<{
+					usersWaitingForE2EKeys: Record<IRoom['_id'], { _id: IUser['_id']; public_key: string }[]>;
+				}>({
+					type: 'object',
+					properties: {
+						usersWaitingForE2EKeys: {
+							type: 'object',
+							additionalProperties: {
+								type: 'array',
+								items: {
+									type: 'object',
+									properties: {
+										_id: { type: 'string' },
+										public_key: { type: 'string' },
+									},
+									required: ['_id', 'public_key'],
+								},
+							},
+						},
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['usersWaitingForE2EKeys', 'success'],
+				}),
+			},
+		},
+
+		async function action() {
+			if (!settings.get('E2E_Enable')) {
+				return API.v1.success({ usersWaitingForE2EKeys: {} });
+			}
+
+			const { roomIds = [] } = this.queryParams;
+			const usersWaitingForE2EKeys = (await Subscriptions.findUsersWithPublicE2EKeyByRids(roomIds, this.userId).toArray()).reduce<
+				Record<string, { _id: string; public_key: string }[]>
+			>((acc, { rid, users }) => ({ [rid]: users, ...acc }), {});
+
+			return API.v1.success({
+				usersWaitingForE2EKeys,
+			});
+		},
+	)
+
+	.post(
 		'e2e.updateGroupKey',
 		{
 			authRequired: true,
@@ -213,6 +392,80 @@ const e2eEndpoints = API.v1
 			await updateGroupKey(rid, uid, key, this.userId);
 
 			return API.v1.success();
+		},
+	)
+	.post(
+		'e2e.provideUsersSuggestedGroupKeys',
+		{
+			authRequired: true,
+			body: isE2EProvideUsersGroupKeyProps,
+			response: {
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+				200: ajv.compile<void>({
+					type: 'object',
+					properties: {
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['success'],
+				}),
+			},
+		},
+
+		async function action() {
+			if (!settings.get('E2E_Enable')) {
+				return API.v1.success();
+			}
+
+			await provideUsersSuggestedGroupKeys(this.userId, this.bodyParams.usersSuggestedGroupKeys);
+
+			return API.v1.success();
+		},
+	)
+	// This should have permissions
+	.post(
+		'e2e.resetRoomKey',
+		{
+			authRequired: true,
+			body: isE2EResetRoomKeyProps,
+			response: {
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+				403: validateForbiddenErrorResponse,
+				200: ajv.compile<void>({
+					type: 'object',
+					properties: {
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['success'],
+				}),
+			},
+		},
+
+		async function action() {
+			const { rid, e2eKey, e2eKeyId } = this.bodyParams;
+			if (!(await hasPermissionAsync(this.userId, 'toggle-room-e2e-encryption', rid))) {
+				return API.v1.forbidden('error-not-allowed');
+			}
+			if (LockMap.has(rid)) {
+				throw new Error('error-e2e-key-reset-in-progress');
+			}
+
+			LockMap.set(rid, true);
+
+			if (!(await canAccessRoomIdAsync(rid, this.userId))) {
+				throw new Error('error-not-allowed');
+			}
+
+			try {
+				await resetRoomKey(rid, this.userId, e2eKey, e2eKeyId);
+				return API.v1.success();
+			} catch (e) {
+				console.error(e);
+				return API.v1.failure('error-e2e-key-reset-failed');
+			} finally {
+				LockMap.delete(rid);
+			}
 		},
 	);
 
@@ -271,117 +524,7 @@ API.v1.addRoute(
 	},
 );
 
-API.v1.addRoute(
-	'e2e.acceptSuggestedGroupKey',
-	{
-		authRequired: true,
-		validateParams: ise2eGetUsersOfRoomWithoutKeyParamsGET,
-	},
-	{
-		async post() {
-			const { rid } = this.bodyParams;
-
-			await handleSuggestedGroupKey('accept', rid, this.userId, 'e2e.acceptSuggestedGroupKey');
-
-			return API.v1.success();
-		},
-	},
-);
-
-API.v1.addRoute(
-	'e2e.rejectSuggestedGroupKey',
-	{
-		authRequired: true,
-		validateParams: ise2eGetUsersOfRoomWithoutKeyParamsGET,
-	},
-	{
-		async post() {
-			const { rid } = this.bodyParams;
-
-			await handleSuggestedGroupKey('reject', rid, this.userId, 'e2e.rejectSuggestedGroupKey');
-
-			return API.v1.success();
-		},
-	},
-);
-
-API.v1.addRoute(
-	'e2e.fetchUsersWaitingForGroupKey',
-	{
-		authRequired: true,
-		validateParams: isE2EFetchUsersWaitingForGroupKeyProps,
-	},
-	{
-		async get() {
-			if (!settings.get('E2E_Enable')) {
-				return API.v1.success({ usersWaitingForE2EKeys: {} });
-			}
-
-			const { roomIds = [] } = this.queryParams;
-			const usersWaitingForE2EKeys = (await Subscriptions.findUsersWithPublicE2EKeyByRids(roomIds, this.userId).toArray()).reduce<
-				Record<string, { _id: string; public_key: string }[]>
-			>((acc, { rid, users }) => ({ [rid]: users, ...acc }), {});
-
-			return API.v1.success({
-				usersWaitingForE2EKeys,
-			});
-		},
-	},
-);
-
-API.v1.addRoute(
-	'e2e.provideUsersSuggestedGroupKeys',
-	{
-		authRequired: true,
-		validateParams: isE2EProvideUsersGroupKeyProps,
-	},
-	{
-		async post() {
-			if (!settings.get('E2E_Enable')) {
-				return API.v1.success();
-			}
-
-			await provideUsersSuggestedGroupKeys(this.userId, this.bodyParams.usersSuggestedGroupKeys);
-
-			return API.v1.success();
-		},
-	},
-);
-
-// This should have permissions
-API.v1.addRoute(
-	'e2e.resetRoomKey',
-	{ authRequired: true, validateParams: isE2EResetRoomKeyProps },
-	{
-		async post() {
-			const { rid, e2eKey, e2eKeyId } = this.bodyParams;
-			if (!(await hasPermissionAsync(this.userId, 'toggle-room-e2e-encryption', rid))) {
-				return API.v1.forbidden();
-			}
-			if (LockMap.has(rid)) {
-				throw new Error('error-e2e-key-reset-in-progress');
-			}
-
-			LockMap.set(rid, true);
-
-			if (!(await canAccessRoomIdAsync(rid, this.userId))) {
-				throw new Error('error-not-allowed');
-			}
-
-			try {
-				await resetRoomKey(rid, this.userId, e2eKey, e2eKeyId);
-				return API.v1.success();
-			} catch (e) {
-				console.error(e);
-				return API.v1.failure('error-e2e-key-reset-failed');
-			} finally {
-				LockMap.delete(rid);
-			}
-		},
-	},
-);
-
-export type E2eEndpoints = ExtractRoutesFromAPI<typeof e2eEndpoints>;
+type E2eEndpoints = ExtractRoutesFromAPI<typeof e2eEndpoints>;
 
 declare module '@rocket.chat/rest-typings' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention, @typescript-eslint/no-empty-interface
