+++ packages/rest-typings/src/v1/e2e.ts
@@ -1,5 +1,3 @@
-import type { IRoom, IUser, ISubscription } from '@rocket.chat/core-typings';
-
 import { ajv } from './Ajv';
 
 type E2eSetUserPublicAndPrivateKeysProps = {
@@ -24,121 +22,8 @@ const E2eSetUserPublicAndPrivateKeysSchema = {
 
 export const isE2eSetUserPublicAndPrivateKeysProps = ajv.compile<E2eSetUserPublicAndPrivateKeysProps>(E2eSetUserPublicAndPrivateKeysSchema);
 
-type E2eGetUsersOfRoomWithoutKeyProps = { rid: string };
-
-const E2eGetUsersOfRoomWithoutKeySchema = {
-	type: 'object',
-	properties: {
-		rid: {
-			type: 'string',
-		},
-	},
-	required: ['rid'],
-	additionalProperties: false,
-};
-
-export const isE2eGetUsersOfRoomWithoutKeyProps = ajv.compile<E2eGetUsersOfRoomWithoutKeyProps>(E2eGetUsersOfRoomWithoutKeySchema);
-
-type E2EProvideUsersGroupKeyProps = {
-	usersSuggestedGroupKeys: Record<IRoom['_id'], { _id: IUser['_id']; key: string; oldKeys: ISubscription['suggestedOldRoomKeys'] }[]>;
-};
-
-const E2EProvideUsersGroupKeySchema = {
-	type: 'object',
-	properties: {
-		usersSuggestedGroupKeys: {
-			type: 'object',
-			additionalProperties: {
-				type: 'array',
-				items: {
-					type: 'object',
-					properties: {
-						_id: { type: 'string' },
-						key: { type: 'string' },
-						oldKeys: {
-							type: 'array',
-							items: {
-								type: 'object',
-								properties: { e2eKeyId: { type: 'string' }, ts: { type: 'string' }, E2EKey: { type: 'string' } },
-							},
-						},
-					},
-					required: ['_id', 'key'],
-					additionalProperties: false,
-				},
-			},
-		},
-	},
-	required: ['usersSuggestedGroupKeys'],
-	additionalProperties: false,
-};
-
-export const isE2EProvideUsersGroupKeyProps = ajv.compile<E2EProvideUsersGroupKeyProps>(E2EProvideUsersGroupKeySchema);
-
-type E2EFetchUsersWaitingForGroupKeyProps = { roomIds: string[] };
-
-const E2EFetchUsersWaitingForGroupKeySchema = {
-	type: 'object',
-	properties: {
-		roomIds: {
-			type: 'array',
-			items: {
-				type: 'string',
-			},
-		},
-	},
-	required: ['roomIds'],
-	additionalProperties: false,
-};
-
-export const isE2EFetchUsersWaitingForGroupKeyProps = ajv.compile<E2EFetchUsersWaitingForGroupKeyProps>(
-	E2EFetchUsersWaitingForGroupKeySchema,
-);
-
-type E2EResetRoomKeyProps = {
-	rid: string;
-	e2eKey: string;
-	e2eKeyId: string;
-};
-
-const E2EResetRoomKeySchema = {
-	type: 'object',
-	properties: {
-		rid: {
-			type: 'string',
-		},
-		e2eKey: {
-			type: 'string',
-		},
-		e2eKeyId: {
-			type: 'string',
-		},
-	},
-	required: ['rid', 'e2eKey', 'e2eKeyId'],
-	additionalProperties: false,
-};
-
-export const isE2EResetRoomKeyProps = ajv.compile<E2EResetRoomKeyProps>(E2EResetRoomKeySchema);
-
 export type E2eEndpoints = {
 	'/v1/e2e.setUserPublicAndPrivateKeys': {
 		POST: (params: E2eSetUserPublicAndPrivateKeysProps) => void;
-	};
-	'/v1/e2e.acceptSuggestedGroupKey': {
-		POST: (params: E2eGetUsersOfRoomWithoutKeyProps) => void;
-	};
-	'/v1/e2e.rejectSuggestedGroupKey': {
-		POST: (params: E2eGetUsersOfRoomWithoutKeyProps) => void;
-	};
-	'/v1/e2e.fetchUsersWaitingForGroupKey': {
-		GET: (params: E2EFetchUsersWaitingForGroupKeyProps) => {
-			usersWaitingForE2EKeys: Record<IRoom['_id'], { _id: IUser['_id']; public_key: string }[]>;
-		};
-	};
-	'/v1/e2e.provideUsersSuggestedGroupKeys': {
-		POST: (params: E2EProvideUsersGroupKeyProps) => void;
-	};
-	'/v1/e2e.resetRoomKey': {
-		POST: (params: E2EResetRoomKeyProps) => void;
 	};
 };
